+++ pkg/cli/cmd/build.go
@@ -10,7 +10,6 @@ type BuildFlags struct {
 	DefinitionFile string
 	ConfigDir      string
 	RootBuildDir   string
-	Validate       bool
 }
 
 var BuildArgs BuildFlags
@@ -22,26 +21,12 @@ func NewBuildCommand(action func(*cli.Context) error) *cli.Command {
 		UsageText: fmt.Sprintf("%s build [OPTIONS]", appName),
 		Action:    action,
 		Flags: []cli.Flag{
-			&cli.StringFlag{
-				Name:        "definition-file",
-				Usage:       "Name of the image definition file",
-				Destination: &BuildArgs.DefinitionFile,
-			},
-			&cli.StringFlag{
-				Name:        "config-dir",
-				Usage:       "Full path to the image configuration directory",
-				Value:       "/eib",
-				Destination: &BuildArgs.ConfigDir,
-			},
+			DefinitionFileFlag,
+			ConfigDirFlag,
 			&cli.StringFlag{
 				Name:        "build-dir",
 				Usage:       "Full path to the directory to store build artifacts",
 				Destination: &BuildArgs.RootBuildDir,
-			},
-			&cli.BoolFlag{
-				Name:        "validate",
-				Usage:       "If specified, the image definition will be validated but not built",
-				Destination: &BuildArgs.Validate,
 			},
 		},
 	}